Extract existing bit access methods to own module
