+++ bst/src/main.rs
@@ -20,6 +20,7 @@
 extern crate alloc;
 
 mod bitcoin;
+mod bits;
 mod characters;
 mod clipboard;
 mod console_out;
